Next version
This upgrade addresses several concerns:

How should transition classes work? What is the API and implementation?
How do we future-proof transition classes for theming?
How can we combine hover and focus in a way that is semantic?
How can we add support for active in a way that is also semantic?
How can we author semantic duration or timing function classes that are also theme aware?